+++ plugins/modules/zhmc_crypto_attachment.py
@@ -54,8 +54,7 @@ options:
     required: true
   hmc_auth:
     description:
-      - The authentication credentials for the HMC, as a dictionary of
-        C(userid), C(password).
+      - The authentication credentials for the HMC.
     type: dict
     required: true
     suboptions:
@@ -69,6 +68,26 @@ options:
           - The password for authenticating with the HMC.
         type: str
         required: true
+      ca_certs:
+        description:
+          - Path name of certificate file or certificate directory to be used
+            for verifying the HMC certificate. If null (default), the path name
+            in the 'REQUESTS_CA_BUNDLE' environment variable or the path name
+            in the 'CURL_CA_BUNDLE' environment variable is used, or if neither
+            of these variables is set, the certificates in the Mozilla CA
+            Certificate List provided by the 'certifi' Python package are used
+            for verifying the HMC certificate.
+        type: str
+        required: false
+        default: null
+      verify:
+        description:
+          - If True (default), verify the HMC certificate as specified in the
+            C(ca_certs) parameter. If False, ignore what is specified in the
+            C(ca_certs) parameter and do not verify the HMC certificate.
+        type: bool
+        required: false
+        default: true
   cpc_name:
     description:
       - The name of the CPC that has the partition and the crypto adapters.
@@ -466,7 +485,7 @@ def ensure_attached(params, check_mode):
 
     # Note: Defaults specified in argument_spec will be set in params dict
     host = params['hmc_host']
-    userid, password = get_hmc_auth(params['hmc_auth'])
+    userid, password, ca_certs, verify = get_hmc_auth(params['hmc_auth'])
     cpc_name = params['cpc_name']
     partition_name = params['partition_name']
     adapter_count = params['adapter_count']
@@ -496,7 +515,8 @@ def ensure_attached(params, check_mode):
     result_changes = dict()
 
     try:
-        session = get_session(_faked_session, host, userid, password)
+        session = get_session(_faked_session,
+                              host, userid, password, ca_certs, verify)
         client = zhmcclient.Client(session)
         cpc = client.cpcs.find(name=cpc_name)
         partition = cpc.partitions.find(name=partition_name)
@@ -901,7 +921,7 @@ def ensure_detached(params, check_mode):
 
     # Note: Defaults specified in argument_spec will be set in params dict
     host = params['hmc_host']
-    userid, password = get_hmc_auth(params['hmc_auth'])
+    userid, password, ca_certs, verify = get_hmc_auth(params['hmc_auth'])
     cpc_name = params['cpc_name']
     partition_name = params['partition_name']
     _faked_session = params.get('_faked_session', None)  # No default specified
@@ -911,7 +931,8 @@ def ensure_detached(params, check_mode):
     result_changes = dict()
 
     try:
-        session = get_session(_faked_session, host, userid, password)
+        session = get_session(_faked_session,
+                              host, userid, password, ca_certs, verify)
         client = zhmcclient.Client(session)
         cpc = client.cpcs.find(name=cpc_name)
         partition = cpc.partitions.find(name=partition_name)
@@ -990,13 +1011,14 @@ def facts(params, check_mode):
     """
 
     host = params['hmc_host']
-    userid, password = get_hmc_auth(params['hmc_auth'])
+    userid, password, ca_certs, verify = get_hmc_auth(params['hmc_auth'])
     cpc_name = params['cpc_name']
     partition_name = params['partition_name']
     _faked_session = params.get('_faked_session', None)  # No default specified
 
     try:
-        session = get_session(_faked_session, host, userid, password)
+        session = get_session(_faked_session,
+                              host, userid, password, ca_certs, verify)
         client = zhmcclient.Client(session)
         cpc = client.cpcs.find(name=cpc_name)
         partition = cpc.partitions.find(name=partition_name)
@@ -1043,7 +1065,16 @@ def main():
     # description of the options in the DOCUMENTATION string.
     argument_spec = dict(
         hmc_host=dict(required=True, type='str'),
-        hmc_auth=dict(required=True, type='dict', no_log=True),
+        hmc_auth=dict(
+            required=True,
+            type='dict',
+            options=dict(
+                userid=dict(required=True, type='str'),
+                password=dict(required=True, type='str', no_log=True),
+                ca_certs=dict(required=False, type='str', default=None),
+                verify=dict(required=False, type='bool', default=True),
+            ),
+        ),
         cpc_name=dict(required=True, type='str'),
         partition_name=dict(required=True, type='str'),
         state=dict(required=True, type='str',